chore: add data-utils import
